ref: rename to layout.wasm and input.wasm
