+++ tasks/bundle-wasm.ts
@@ -27,7 +27,11 @@ function encodeZ85(data: Uint8Array): string {
   return out.join("");
 }
 
-const [input = "clayterm.wasm", output = "wasm.ts"] = Deno.args;
+const [input, output] = Deno.args;
+if (!input || !output) {
+  console.error("Usage: bundle-wasm.ts <input.wasm> <output.ts>");
+  Deno.exit(1);
+}
 
 const wasm = await Deno.readFile(input);
